Max Pairwise Product
Maximum Pairwise Product Problem
Find the maximum product of two distinct numbers in a sequence of non-negative integers.
Input: A sequence of non-negative
integers.
Output: The maximum value that
can be obtained by multiplying
two different elements from the sequence.
5 6 2 7 4
5 30 10 35 20
6 30 12 42 24
2 10 12 7 4
7 35 42 14 28
4 20 24 8 28
Given a sequence of non-negative integers a
1
;:::;a
n
, compute
max
1i ,j n
a
i
 a
j
:
Note that i and j should be different, though it may be the case that a
i
= a
j
.
Input format. The first line contains an integer n. The next line contains
n non-negative integers a
1
;:::;a
n
(separated by spaces).
Output format. The maximum pairwise product.
Constraints. 2  n  2  10
5
; 0  a
1
;:::;a
n
 2  10
5
.
Sample 1.
Input:
3
1 2 3
Output:
6
Sample 2.
Input:
10
7 5 14 2 8 8 10 1 2 3
Output:
140
Time and memory limits. The same as for the previous problem.
2.1 Naive Algorithm
A naive way to solve the Maximum Pairwise Product Problem is to go
through all possible pairs of the input elements A[1 :::n] = [a
1
;:::;a
n
] and
to find a pair of distinct elements with the largest product:
MaxPairwiseProductNaive(A[1 :::n]):
product   0
for i from 1 to n:
for j from 1 to n:
if i , j :
if product < A[i ]  A[j ]:
product   A[i ]  A[j ]
return product
This code can be optimized and made more compact as follows.
MaxPairwiseProductNaive(A[1 :::n]):
product   0
for i from 1 to n:
for j from i + 1 to n:
product   max(product;A[i ]  A[j ])
return product
Implement this algorithm in your favorite programming language. If
you are using C++, Java, or Python3, you may want to download the starter
files (we provide starter solutions in these three languages for all the problems in the book). For other languages, you need to implement your solution from scratch.